db/init_vectors.py
import json
import logging
import numpy as np
import faiss
import os
from pathlib import Path
from sqlalchemy.orm import Session

logger = logging.getLogger(__name__)

# Directory to store vector indices
VECTOR_DIR = Path("./vector_indices")

def init_vector_db() -> None:
    """Initialize the vector database for content embeddings."""
    VECTOR_DIR.mkdir(exist_ok=True)
    
    # Check if index already exists
    index_path = VECTOR_DIR / "content_embeddings.index"
    
    if index_path.exists():
        logger.info("Vector index already exists. Skipping initialization.")
        return
    
    # Create a simple vector index
    # For the prototype, we'll use random vectors - in production this would use 
    # actual embeddings from a model like sentence-transformers
    dimension = 128  # embedding dimension
    index = faiss.IndexFlatL2(dimension)
    
    # Save the empty index
    faiss.write_index(index, str(index_path))
    logger.info(f"Created new vector index at {index_path}")

def update_content_embeddings(db: Session) -> None:
    """Update vector embeddings for all curriculum content."""
    # In a real implementation, this would load content from the database
    # and generate embeddings using a model
    
    # For the prototype, we'll just check if we need to update
    index_path = VECTOR_DIR / "content_embeddings.index"
    if not index_path.exists():
        logger.error("Vector index not found. Call init_vector_db first.")
        return
    
    # Load the index
    index = faiss.read_index(str(index_path))
    
    # In a real implementation, we would:
    # 1. Get all content from the database
    # 2. Generate embeddings for each content
    # 3. Add the embeddings to the index
    
    # For the prototype, we'll just log that we would update
    logger.info("Would update content embeddings (placeholder - no real action taken)")
    
    # Save the index with metadata
    metadata_path = VECTOR_DIR / "content_embeddings.json"
    with open(metadata_path, "w") as f:
        json.dump({
            "last_updated": str(os.path.getmtime(str(index_path))),
            "dimension": index.d,
            "size": index.ntotal
        }, f)
    
    logger.info("Content embeddings metadata updated")
