requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pydantic>=2.6.0
email-validator>=2.1.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
pymongo>=4.6.0
redis>=5.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# ML & Data Science
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0
torch>=2.1.0
transformers>=4.36.0
sentence-transformers>=2.2.0

# Vector Storage
faiss-cpu>=1.7.4
chromadb>=0.4.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
httpx>=0.26.0

# Development Tools
black>=24.0.0
flake8>=7.0.0
isort>=5.13.0
mypy>=1.8.0

# Monitoring & Logging
prometheus-client>=0.19.0
python-json-logger>=2.0.7

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.5.0

# Utilities
python-dateutil>=2.8.2
pytz>=2024.1
tqdm>=4.66.0 